pyproject.toml
[project]
name = "airline-disruption-env"
version = "1.0.0"
description = "Airline disruption recovery environment for evaluating AI agents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pyyaml",
    "openai",
    "requests",
    "python-dotenv",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.openenv]
name = "airline-disruption-env"
entrypoint = "app.py"
